[Feat] : 감사서클 user count 적용
