+++ components/screen/ParticipantNickname.tsx
@@ -7,7 +7,6 @@ import { usePublish, useDisconnect } from '@/atoms/socketAtoms';
 import userNickname from '@/atoms/userNickname';
 import useNavigation from '@/hooks/useNavigation';
 import useSocket from '@/hooks/useSocket';
-import useUserCount from '@/hooks/useUserCount';
 import localStorage from '@/service/localStorage';
 
 import ElButton from '../elements/ElButton';
@@ -25,8 +24,6 @@ export default function ScreenParticipantNickname() {
 
 	const { connect, publish, disconnect } = useSocket();
 
-	const { connect: connectCount } = useUserCount();
-
 	const handleNickname = (e: React.ChangeEvent<HTMLInputElement>) => {
 		setNickname(e.target.value);
 	};
@@ -40,8 +37,6 @@ export default function ScreenParticipantNickname() {
 
 		setPublish(() => publish);
 		setDisconnect(() => disconnect);
-
-		connectCount(`/app/${roomName[1]}/${roomName[2]}/get-info`, nickname);
 
 		switch (roomName[1]) {
 			case 'speedgame':
